Extract `WebDavUrlChecker` from `WebDavMountRepository`
