+++ core/src/androidTest/kotlin/at/bitfire/davdroid/webdav/WebDavUrlCheckerTest.kt
@@ -17,13 +17,13 @@ import org.junit.Test
 import javax.inject.Inject
 
 @HiltAndroidTest
-class WebDavMountRepositoryTest {
+class WebDavUrlCheckerTest {
 
     @get:Rule
     val hiltRule = HiltAndroidRule(this)
 
     @Inject
-    lateinit var repository: WebDavMountRepository
+    lateinit var webDavUrlChecker: WebDavUrlChecker
 
     @Before
     fun setUp() {
@@ -34,33 +34,44 @@ class WebDavMountRepositoryTest {
     val url = web.url("/")
 
     @Test
-    fun testHasWebDav_NoDavHeader() = runTest {
+    fun getWebDavUrl_NoDavHeader() = runTest {
         web.enqueue(MockResponse().setResponseCode(200))
-        assertNull(repository.hasWebDav(url, null))
+
+        val result = webDavUrlChecker.getWebDavUrl(url = url, credentials = null)
+
+        assertNull(result)
     }
 
     @Test
-    fun testHasWebDav_DavClass1() = runTest {
+    fun getWebDavUrl_DavClass1() = runTest {
         web.enqueue(MockResponse()
             .setResponseCode(200)
             .addHeader("DAV: 1"))
-        assertEquals(url, repository.hasWebDav(url, null))
+
+        val result = webDavUrlChecker.getWebDavUrl(url = url, credentials = null)
+
+        assertEquals(url, result)
     }
 
     @Test
-    fun testHasWebDav_DavClass2() = runTest {
+    fun getWebDavUrl_DavClass2() = runTest {
         web.enqueue(MockResponse()
             .setResponseCode(200)
             .addHeader("DAV: 1, 2"))
-        assertEquals(url,repository.hasWebDav(url, null))
+
+        val result = webDavUrlChecker.getWebDavUrl(url = url, credentials = null)
+
+        assertEquals(url, result)
     }
 
     @Test
-    fun testHasWebDav_DavClass3() = runTest {
+    fun getWebDavUrl_DavClass3() = runTest {
         web.enqueue(MockResponse()
             .setResponseCode(200)
             .addHeader("DAV: 1, 3"))
-        assertEquals(url,repository.hasWebDav(url, null))
-    }
 
+        val result = webDavUrlChecker.getWebDavUrl(url = url, credentials = null)
+
+        assertEquals(url, result)
+    }
 }
